+++ ui/cypress/e2e/explorer.test.ts
@@ -351,6 +351,10 @@ describe('DataExplorer', () => {
   describe('raw script editing', () => {
     beforeEach(() => {
       cy.getByTestID('switch-to-script-editor').click()
+
+      // explicitly wait, to give the CodeMirror editor a chance to render
+      // may not be necessary when we switch to Monaco
+      cy.wait(300)
     })
 
     it('enables the submit button when a query is typed', () => {